Replaced 'float' with 'double'; removed old broken stuff

+++ src/Buddhabrot/Constant.cs
@@ -7,7 +7,7 @@ namespace Buddhabrot
         // The upper range is hard-coded in the OpenCL kernel
         public static readonly IntRange IterationRange = new IntRange(20_000_000, 30_000_000);
         public static readonly ComplexArea RenderingArea = new ComplexArea(
-            realRange: new FloatRange(-1.45f, 0.75f),
-            imagRange: new FloatRange(-1.1f, 1.1f));
+            realRange: new DoubleRange(-1.45f, 0.75f),
+            imagRange: new DoubleRange(-1.1f, 1.1f));
     }
 }
